Update some type extensions
