+++ CommunityToolkit.Mvvm.SourceGenerators/Extensions/INamedTypeSymbolExtensions.cs
@@ -2,9 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections.Generic;
-using CommunityToolkit.Mvvm.SourceGenerators.Helpers;
 using Microsoft.CodeAnalysis;
 
 namespace CommunityToolkit.Mvvm.SourceGenerators.Extensions;
@@ -14,55 +12,6 @@ namespace CommunityToolkit.Mvvm.SourceGenerators.Extensions;
 /// </summary>
 internal static class INamedTypeSymbolExtensions
 {
-    /// <summary>
-    /// Gets a fully qualified metadata name for a given <see cref="INamedTypeSymbol"/> instance.
-    /// </summary>
-    /// <param name="symbol">The input <see cref="INamedTypeSymbol"/> instance.</param>
-    /// <returns>The fully qualified metadata name for <paramref name="symbol"/>.</returns>
-    public static string GetFullyQualifiedMetadataName(this INamedTypeSymbol symbol)
-    {
-        using ImmutableArrayBuilder<char> builder = ImmutableArrayBuilder<char>.Rent();
-
-        static void BuildFrom(ISymbol? symbol, in ImmutableArrayBuilder<char> builder)
-        {
-            switch (symbol)
-            {
-                // Namespaces that are nested also append a leading '.'
-                case INamespaceSymbol { ContainingNamespace.IsGlobalNamespace: false }:
-                    BuildFrom(symbol.ContainingNamespace, in builder);
-                    builder.Add('.');
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-                // Other namespaces (ie. the one right before global) skip the leading '.'
-                case INamespaceSymbol { IsGlobalNamespace: false }:
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-                // Types with no namespace just have their metadata name directly written
-                case ITypeSymbol { ContainingSymbol: INamespaceSymbol { IsGlobalNamespace: true } }:
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-                // Types with a containing non-global namespace also append a leading '.'
-                case ITypeSymbol { ContainingSymbol: INamespaceSymbol namespaceSymbol }:
-                    BuildFrom(namespaceSymbol, in builder);
-                    builder.Add('.');
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-                // Nested types append a leading '+'
-                case ITypeSymbol { ContainingSymbol: ITypeSymbol typeSymbol }:
-                    BuildFrom(typeSymbol, in builder);
-                    builder.Add('+');
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        BuildFrom(symbol, in builder);
-
-        return builder.ToString();
-    }
-
     /// <summary>
     /// Gets all member symbols from a given <see cref="INamedTypeSymbol"/> instance, including inherited ones.
     /// </summary>
